Move v2 backend to subfolders and mount controllers
