+++ server/model/v2/SplusEinsModel.ts
