+++ server/model/v2/SpluseEinsEvent.spec.ts
