--- server_2.0/README.md
@@ -1,16 +1,0 @@
-# SERVER / BACKEND
-
-This directory contains the SplusEins backend files.
-
-The backend is an Express.js application.
-
-More information about the usage of this directory in [the documentation](https://expressjs.com/).
-
-**Base Endpoints**:
-
-  * /api/splus
-  * /api/ics
-  * /api/mensa
-  * /api/news
-
-Detailed information about the endpoints are in the specific controller files.
