--- server_2.0/index.ts
@@ -1,9 +1,0 @@
-import App from './App';
-
-const host = process.env.HOST || '127.0.0.1';
-const port = parseInt(process.env.PORT) || 3001;
-const path = process.env.API_PREFIX;
-
-new App(path).app.listen(port, host, () => {
-  console.log(`listening on host ${host} and port ${port}`)
-});
