--- server_2.0/serverless/api.ts
@@ -1,8 +1,0 @@
-import * as awsServerlessExpress from 'aws-serverless-express';
-import App from '../App';
-
-const server = awsServerlessExpress.createServer(new App('/.netlify/functions/api').app); // TODO make this configurable
-
-export const handler = (event, context) => {
-  awsServerlessExpress.proxy(server, event, context);
-};
